USB: phy: Disable IRQs before resetting clocks
As a part of msm_otg_reset, we go for disabling and
enabling of clocks. There is a chance that IRQ may
get fired when the clocks are disabled. This may lead
to NOC error because of unclocked access when we try
to access some memory region as a part of IRQ handling.

Fix this by disabling IRQs before going for clock reset
and enable IRQs later.

Change-Id: I8ccec08c9c520f0573a2dbe1a9f8e780b3516e85
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: Sai Krishna Juturi <jsaikrishna@codeaurora.org>