wlan: Parse PNO channel number from FW
Host calculates the channel number for the Preferred
network BSSIDs from the received frame. If the channel
info IEs are not present in the beacon frame, then channel
determination fails in the host. Host will assign default
channel 1 or 36 based on the band.

If the channel number is not available in beacon, get the
channel number from the firmware and use the same.

Change-Id: If6abe81732ea671ee99a4f111329567a2a4f0a87
CRs-Fixed: 3097913